+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Fix DI issues in ASP.NET Core + SentryHttpMessageHandlerBuilderFilter (#789) @Tyrrrz
+- Fix incorrect NRT on SpanContext.ctor (#788) @Tyrrrz
 
 ## 3.0.2